+++ packages/badge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/badge
 
+## 0.7.4
+
+### Patch Changes
+
+- @inlang/sdk@0.26.5
+
+## 0.7.3
+
+### Patch Changes
+
+- @inlang/sdk@0.26.4
+
 ## 0.7.2
 
 ### Patch Changes