+++ packages/github-lint-action/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/github-lint-action
 
+## 0.1.6
+
+### Patch Changes
+
+- @inlang/sdk@0.26.5
+
+## 0.1.5
+
+### Patch Changes
+
+- @inlang/sdk@0.26.4
+
 ## 0.1.4
 
 ### Patch Changes